refactor: reorganize socket code into logical subdirectories
- Move actors into actors/ subdirectory:
  - chat.py → actors/assistant.py
  - base.py → actors/base.py
  - claude_sdk_actor.py → actors/claude_sdk.py

- Move types into types/ subdirectory:
  - envelope_type.py → types/envelope.py
  - types.py → types/message.py

- Move utilities into utils/ subdirectory:
  - streamer.py → utils/streamer.py

- Rename setup.py → handlers.py for clarity

- Update all imports across the codebase to use new paths

- Remove claude.py (handler moved to actors/claude_sdk.py)

This improves code organization by grouping related files together,
making the codebase easier to navigate and maintain.
