+++ backend/core/sockets/actors/__init__.py
@@ -1,0 +1,11 @@
+from .assistant import AssistantActor, handle_chat_stream_start
+from .base import BaseActor
+from .claude_sdk import ClaudeSDKActor, request_claude_stream
+
+__all__ = [
+    "BaseActor",
+    "AssistantActor",
+    "ClaudeSDKActor",
+    "handle_chat_stream_start",
+    "request_claude_stream",
+]
