+++ backend/core/sockets/utils/__init__.py
@@ -1,0 +1,3 @@
+from .streamer import stream_chunks_openai
+
+__all__ = ["stream_chunks_openai"]
